Refactor project structure following image-search pattern
- Move Python source code to src/ directory
  - src/agents/: Agent related modules
  - src/common/: Common utilities (connection)
  - src/frontend/: UI components (chat_ui, flow_ui, i18n)
  - src/rag/: RAG related modules
  - src/tools/: Tool scripts (embed_docs, convert_headings, etc.)
- Move shell scripts to scripts/ directory
- Create data/ directory for doc_repos and uploaded files
- Update all import paths
- Add start.sh script for easy UI launching
- Update README.md and README_zh.md with new paths

Resolves #2
