+++ src/tools/prepare_bgem3.py
